+++ packages/fluekit/src/index.ts
@@ -79,6 +79,7 @@ export { default as CircleAvatar } from "./CircleAvatar.vue";
 export { default as TabBar } from "./TabBar.vue";
 export { default as TabBarView } from "./TabBarView.vue";
 export { default as RefreshIndicator } from "./RefreshIndicator.vue";
+export { default as RatingBar } from "./RatingBar.vue";
 export { default as FlueConfigProvider } from "./FlueConfigProvider.vue";
 export * from "./Color";
 export * from "./ImageUtils";